Implement updatePolicy function
The function takes a startDays and endDays argument to indicate the 
range of booking lengths for which to apply the discount. If the 
provider wishes to apply a 3% discount for all bookings between 3 days 
and 5 days long (inclusive) then startDays=3 and endDays=5 and 
discount=0.3. 

startDays should never be 0. The indices used to indicate the booking 
length is shifted by one to start from 0.