Add exception to getPricing
The exception catches the case when a getPricing is called using a
bikeType that doesn't exist in the pricing policy data structure.

Th reason this is a runtime error is because there is no way of knowing
what bike types have been added to the pricing policy data structure
until after runtime.

The error is not caught because there is no obvious way for the system
to deal with such a situation. The system should prevent the providers
from renting out bikes without specifying daily rental prices for such
bike types so raising an uncaught exception is reasonable in this case
because it indicates some other part of the system is broken.